add flare finding algorithm options
