+++ laff/__init__.py
@@ -1,3 +1,3 @@
-from .laff import fitGRB, plotGRB, set_logging_level
+from .laff import fitGRB, plotGRB, set_logging_level, findFlares, fitContinuum, fitFlares
 from .lightcurve_import import lcimport
 from .utility import STAT_NAMES_CONTINUUM, PAR_NAMES_FLARE, PAR_NAMES_CONTINUUM
