+++ laff/flarefinding/algorithms/__init__.py
@@ -1,0 +1,1 @@
+from .sequential_v1 import sequential_findflares
