+++ laff/flarefinding/algorithms/savitsky.py
@@ -1,0 +1,11 @@
+
+
+# First find obvious flares
+# Big rises mark as flares
+
+
+# Do smoothing on the data
+# Then do lightcurve minus smoothed data
+# Any peaks - mark as flares
+
+# Repeat a few times?
